Project rebuilt with visual navigation; works but not tested on physical robot.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TurtleDozer.kt
@@ -278,7 +278,7 @@ class TurtleDozer(
     }
 
     private fun slideOver(inches: Double) {
-        driveByEncoder(Vector(inches.toInt(), 0))
+        driveByEncoder(Vector(inches, 0.0))
     }
 
     fun dragWithBalance(telemetry: Telemetry) {
